Refactor sass into separate files
