+++ src/routes/SignIn/components/Help.js
@@ -2,6 +2,7 @@ import React from 'react';
 import Dialog from 'material-ui/Dialog';
 import FlatButton from 'material-ui/FlatButton';
 import RaisedButton from 'material-ui/RaisedButton';
+import './Help.scss'
 
 
 export default class Help extends React.Component {
