+++ src/routes/SignIn/components/Help.scss
@@ -1,0 +1,5 @@
+.help-header {
+  text-transform: uppercase;
+  font-weight: normal;
+  font-size: 1.1rem;
+}
