+++ src/routes/SignIn/components/SignInPage.scss
@@ -1,0 +1,6 @@
+.sign-in-background {
+  background-image: url("../../../../public/background.jpg");
+  background-attachment: fixed;
+  background-size: cover;
+  background-position: center;
+}
