+++ src/routes/SignIn/index.js
@@ -1,0 +1,5 @@
+import SignInPage from './components/SignInPage.js'
+
+export default {
+  component : SignInPage
+}
