--- src/routes/SignIn/index.jsx
@@ -1,6 +1,0 @@
-import SignInPage from './components/SignInPage.jsx'
-
-export default {
-  component : SignInPage
-
-}
