+++ src/styles/global.scss
@@ -1,0 +1,86 @@
+html {
+  box-sizing: border-box;
+}
+
+html,
+body {
+  color: $text-color;
+  font-family: 'Roboto', sans-serif;
+  margin: 0;
+  padding: 0;
+
+  min-width: 100%;
+  width: 100%;
+  max-width: 100%;
+
+  min-height: 100%;
+  height: 100%;
+  max-height: 100%;
+}
+
+#root > [data-reactroot] {
+    min-height: 100% !important;
+    height: 100% !important;
+}
+
+.root {
+  background: linear-gradient($light-bkgrnd, $bkgrnd-gradient);
+}
+
+*,
+*:before,
+*:after {
+  box-sizing: inherit;
+}
+
+.box-shadow {
+  box-shadow: 2px 2px 2px $shadow-color;
+}
+
+.button-shadow {
+  box-shadow: 0px 0px 2px 2px $text-color !important;
+}
+
+.circle-highlight {
+  box-shadow: 0px 0px 2px 1px $shadow-color;
+}
+
+.overflow-wrap {
+  overflow-wrap: break-word;
+  hyphens: auto;
+}
+
+.circle {
+  border-radius: 100%;
+}
+
+.highlighted-white-text {
+  color: white;
+  text-shadow: 1px 1px 1px black;
+  font-weight: bold;
+}
+
+.show-pointer {
+  cursor: pointer;
+}
+
+.full-width {
+  width: 100%;
+}
+
+a {
+  color: $material-ui-blue;
+}
+
+.max-page-height {
+  height: 100%;
+  min-height: 100%;
+}
+
+.min-page-height {
+  min-height: 100%;
+}
+
+.transparent {
+  background-color: transparent;
+}
